refactor: moved hooks into separate folder.
